streamlit_requirements.txt
streamlit
Pillow
numpy
torch
timm
torchvision
thop
yacs
opencv-python